backend/requirements.txt
fastapi
uvicorn
python-socketio
opencv-python-headless
numpy
requests
python-dotenv
aiohttp